+++ packages/core/src/List.js
@@ -1,61 +1,51 @@
-// @flow
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import type { ReactChildren } from 'react-flow-types';
 import './styles/List.scss';
 
 import prefixClass from './utils/prefixClass';
 import icBEM from './utils/icBEM';
 
+import Section from './Section';
+
 export const COMPONENT_NAME = prefixClass('list');
 const ROOT_BEM = icBEM(COMPONENT_NAME);
 export const BEM = {
     root: ROOT_BEM,
-    title: ROOT_BEM.element('title'),
     body: ROOT_BEM.element('body'),
-    desc: ROOT_BEM.element('desc'),
 };
 
 const NORMAL = 'normal';
-const SETTING = 'setting';
+const SETTING = 'setting'; // #TODO: design deprecated
 const BUTTON = 'button';
 const LIST_VARIANTS = [NORMAL, SETTING, BUTTON];
 
-export type Props = {
-    variant: typeof NORMAL | typeof SETTING | typeof BUTTON,
-    title?: string,
-    desc?: ReactChildren,
-
-    /* eslint-disable react/require-default-props */
-    className?: string,
-    children?: ReactChildren,
-    /* eslint-enable react/require-default-props */
-};
+export const TYPE_SYMBOL = Symbol('List');
 
 function List({
     variant,
+    // <Section> props
     title,
     desc,
     // React props
     className,
     children,
     ...otherProps,
-}: Props) {
+}) {
     const bemClass = BEM.root.modifier(variant);
     const rootClassName = classNames(bemClass.toString(), className);
 
-    const titleNode = <div className={BEM.title.toString()}>{title}</div>;
-    const descNode = <div className={BEM.desc.toString()}>{desc}</div>;
-
     return (
-        <div className={rootClassName} {...otherProps}>
-            {title && titleNode}
+        <Section
+            className={rootClassName}
+            title={title}
+            desc={desc}
+            bodySpacing={false}
+            {...otherProps}>
             <ul className={BEM.body.toString()}>
                 {children}
             </ul>
-            {desc && descNode}
-        </div>
+        </Section>
     );
 }
 
@@ -70,5 +60,9 @@ List.defaultProps = {
     title: undefined,
     desc: undefined,
 };
+
+// For `<ListRow>` to check if `nestedList` is a `<List>.
+// Ref for this symbol approach: https://github.com/iCHEF/gypcrete/pull/157
+List.typeSymbol = TYPE_SYMBOL;
 
 export default List;